+++ core/app_administration.go
@@ -215,3 +215,12 @@ func (app *application) admGetOrganizations(ID string) ([]model.Organization, er
 
 	return getOrganization, nil
 }
+
+func (app *application) admUpdateOrganization(ID string, name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) error {
+	err := app.storage.UpdateOrganization(ID, name, requestType, requiresOwnLogin, loginTypes, organizationDomains)
+	if err != nil {
+		return err
+	}
+
+	return err
+}